Moves handling of scripts into a new class
Keeping it separate from jobs not only makes jobs less polluted but allows scripting to be used outside job contexts if necessary.

Fixes: SIRI-1120
